Refactor Stats Columns to sysystems data file
